requirements.txt
Flask>=2.0
Flask-WTF>=1.1.1
python-dotenv>=0.21.0
Werkzeug>=2.0